egl: Set the attrib_list as *const i32
Original: 7f23dd980dd74affbc7cd0687a7ec9105aedb763

+++ src/backend/egl/display.rs
@@ -111,7 +111,7 @@ fn select_platform_display<N: EGLNativeDisplay + 'static>(
                         ffi::egl::GetPlatformDisplayEXT(
                             platform.platform,
                             platform.native_display,
-                            platform.attrib_list.as_ptr() as *const isize,
+                            platform.attrib_list.as_ptr() as *const i32,
                         )
                     }
                 }
